implements checkpoint function of dcp in POSIX

+++ deps/md5/md5.h
@@ -48,7 +48,7 @@ extern void MD5_Final(unsigned char *result, MD5_CTX *ctx);
 
 #ifndef HAVE_OPENSSL
 // openssl provide a function MD5(), we just wrap it out if the library is not used
-extern unsigned char * MD5( void *pointer, size_t pointerLength, unsigned char *md5HashPointer);
+extern unsigned char * MD5( void *pointer, unsigned long pointerLength, unsigned char *md5HashPointer);
 #endif
 
 #endif
